Changed Layout
To test my HTML and CSS skills, I went ahead and tried to upgrade my first ever website to something more appealing. These are my changes to the web page after about of month of learning HTML and CSS. I've mainly focused on the homepage (index.html), but plan on fixing all other pages connected to this in the future.